Split test graph file-read host effect fixtures
